Remove unnecessary CI support code
`fast_pass` isn't needed in this repository, because it only contains
one crate (effectively; `habitat_win_users` is here, but it's an
implementation detail of `habitat_core` on Windows).

Builds were silently failing because there was an extra helper script
that `fast_pass` needed, but it wasn't carried over in the split.

Also make all Linux CI scripts executable.

Signed-off-by: Christopher Maier <cmaier@chef.io>