[core] Restore impl for PackageInstall#paths() to use `PATH` metafile.
This change restores the previous implementation of the
`PackageInstall#paths()` method to return path entries directly from the
`PATH` metafile and to not use the `RUNTIME_ENVIRONMENT` metafile.

As referenced in habitat-sh/builder#240, the semantic meaning of the
`PATH` metafile is precise and distinct from the newer
`RUNTIME_ENVIRONMENT` metafile. The path entries in a `PATH` metafile
are a one-to-one mapping with the values in the `$pkg_bin_dirs` Plan
variable which created the package. That is, the path entries in `PATH`
should only refer to directories under the package's own
`$pkg_prefix`--not to any other package. There are certain activities
such as searching for a binary in a package, performing binlinking, etc.
which use the `PATH` metafile and addition entires from external
packages break these implementations.

All packages built with a build program after
https://github.com/habitat-sh/habitat/commit/13344a679155e5210dd58ecb9d94654f5ae676d3
which merged in habitat-sh/habitat#4067 (and released in Habitat 0.50.0)
may contain extra path entries in their `PATH` metafiles. To accommodate
these packages, an additional filtering step is performed in the
`#paths()` method to drop all path entries that fall outside of the
package's `$pkg_prefix`. Several net-new unit tests are added with this
change to ensure the correct behavior in future feature work.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>