[api-client] Use `PackageTarget::active_target()` in user agent string.
This change slightly updates the `User-Agent` header derivation which is
used in API requests. Prior to this change the current system's
architecture and system values were pulled from a `uname(1)` call and
thus were runtime dynamic. This change changes that to use the already
available `PackageTarget::active_target()` implementation which is
currently set at build time and is therefore static. The set of valid
Habitat clients to date should mean that there is no net difference
between the two, however there is a chance that the value of `<TARGET>`
in the user agent string may have slightly changed for some clients.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>