Refactor `PackageIdent` to an enum-based representation.
This is an initial refactoring of the current struct-based
`PackageIdent` type to use an enum-based representation, called `Ident`.
For existing compatibility, `Ident` is currently exported under its
former name of `PackageIdent` but this can be updated later. This newer
representation attempts to follow API conventions and idioms closer to
those used in Rust's standard library (i.e. standard API interfaces,
common trait implementations, etc.).

The prior implementation follows a very c-like struct pattern (i.e.
broken out public fields for each part of an identifier) and therefore
needed to add `Option`s around some of the fields. In Habitat, we have 3
legal forms of a package identifier:

* `$origin/$name` - which is called a `NameIdent`
* `$origin/$name/$version` - which is called a `VersionIdent`
* `$origin/$name/$version/$release` - which is called a `ReleaseIdent`

Note that an origin-only identifier is not valid, nor is an identifier
which omits a `$version` but provides a `$release`. This latter case is
permitted in the prior implementation and can result in unexpected
behavior when used and consumed (often resulting in a panic is an
unconditional `.unwrap()` is expected).

The `Ident` type attempts to provide either identical or similarly
appropriate methods in an attempt to be minimally intrusive at existing
call sites. Each enum variant has a proper "subtype" which is home to
each type's implementation when the implementation is varied (for
example, the `std::fmt::Display` implementations). This also means,
importantly, that each proper subtype can be used in Habitat code when a
particular variant is expected and/or required. The most useful case
here is the ability to represent what was formerly referred to as a
"fully qualified ident", which is called a "release ident" in this
change (represented by `ReleaseIdent`). In surveying the existing
codebases, it appears as though release idents are used in the vast
majority of cases whereas name or version idents (formerly referred to
as "fuzzy idents") are typically only used at the edges of a system
(such as user input in CLI parsing) or saved as end-user's intent (such
as the identifier provided when starting a Supervisor service).

There are some transitionary implementations and methods which help
preserve existing behavior (such as retaining the `.fully_qualified()`
method) but are marked with a note for later removal. Each of these
removals will likely result in downstream changes which help tighten up
the calling code but may require more effort (hence starting small and
knowing the direction of work ahead).

Note that some of the refactored code, namely the version and ident
sorting logic needs standalone unit tests and more rigorous testing of
the provided behavior is needed to ensure that backwards compatibility
of behavior is preserved.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>