Run tests in both normal (debug) mode and release mode.
This might seem odd, but we have certain code that behaves different
when in debug mode vs in release mode, and the only way to test both
code paths is to run the tests in both modes. It adds a bit of time to
the test suite, but the test suite is small enough that this is
negligible overall, and gives us a tangible benefit in terms of
robustness.

Signed-off-by: Josh Black <raskchanky@gmail.com>