Add EnvConfig from habitat as env::Config and implement it for ChannelIdent
There are some compromises here since it's not possible to directly reproduce
the previous behavior within the context of the env::Config trait. That is,
the HAB_BLDR_CHANNEL environment variable was used as the primary override
for the ChannelIdent default, but if that was not set, it would fall back to
trying the legacy environment variable: HAB_DEPOT_CHANNEL.

Since env::Config uses just one environment variable to provide a
ChannelIdent::configured_value instance, we can't have the same behavior
without putting environment variable overriding logic into
ChannelIdent::default, which kind of defeats the purpose of the uniform
env::Config approach.

As such, we instead make a clean break with the legacy approach:
- ChannelIdent::default returns the stable channel
- ChannelIdent::configured_value returns a ChannelIdent instance for the
  value in HAB_BLDR_CHANNEL if set and the default value otherwise
- HAB_DEPOT_CHANNEL no longer has any special treatment

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>